Adds check for configuration

+++ src/Command/InstallCommand.php
@@ -2,10 +2,11 @@
 namespace KindlingCLI\Command;
 
 use KindlingCLI\WPCLI\WP;
+use KindlingCLI\Command\PostInstall;
+use KindlingCLI\Option\LocalConfiguration;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use KindlingCLI\Command\PostInstall;
 
 class InstallCommand extends Command
 {
@@ -17,6 +18,14 @@ class InstallCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        // Make sure the configuration exists.
+        if (!LocalConfiguration::configExists()) {
+            $output->writeln(
+                '<error>Configuration file not present run kindling init {project-name} to get started.</error>'
+            );
+            exit(1);
+        }
+
         $output->writeln("<info>== Running Pre-install Commands =======================</info>");
 
         // Run pre-install commands.
